test: 로또 발행 내역 출력 기능 테스트
